+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next Release
 
+## 1.17.3
+
 - Remove "Oracle NN:" from Ironsworn oracle names ([#453](https://github.com/ben/foundry-ironsworn/pull/453))
 - Fix a visual glitch with progress tracks in v10
 - Retire the classic Site sheet